lua now compiles with lua_Number defined as a float with 0 warnings. To get back to normal, double-flavored lua, simply revert all the lua source files to the previous revision and then change the #define back to double.
Added licenses.

drawFPS() now draws all text twice so it looks like a drop-shadow font, but really it isn't.
It also draws "DONGS" once for each excess millisecond between Display() calls, so that's nice.

+++ trunk/lua/src/luaconf.h
@@ -10,6 +10,7 @@
 
 #include <limits.h>
 #include <stddef.h>
+#include <math.h>
 
 
 /*
@@ -534,8 +535,8 @@
 #define luai_numsub(a,b)	((a)-(b))
 #define luai_nummul(a,b)	((a)*(b))
 #define luai_numdiv(a,b)	((a)/(b))
-#define luai_nummod(a,b)	((a) - floor((a)/(b))*(b))
-#define luai_numpow(a,b)	(pow(a,b))
+#define luai_nummod(a,b)	((a) - floorf(((a)/(b)))*(b))
+#define luai_numpow(a,b)	(powf(a,b))
 #define luai_numunm(a)		(-(a))
 #define luai_numeq(a,b)		((a)==(b))
 #define luai_numlt(a,b)		((a)<(b))
